build: fix compile warnings
klee/searcher: remove unused field from filtersearcher
skins/partseedcore: nuke unused exe field
klee/priorirty searcher: remove unused field 'cmp'
klee/kmodule: remove unused modlue field from raiseasm pass
klee/solver: remove unused query_hash field in RandomValue
klee/solver: remove unused solver field in stpsolver
klee/expr: remove unused 'verified' field in rule builder
klee/statesolver/ktest: fix override warnings
statesolver/costkiller: fix override warnings, etc
kopt/knockoutrule: ifdef out unused function to get rid of warnings
klee-mc/vfd: move unused stack variable to fix warning
skin/ddt: fix override warnings
klee/executortimers: remove unused stuff to fix warnings
klee/memorymanager: move default case in switch around to fix warning
kmc-replay/syscallsktest: fix value width/sign warnings
klee-replay: fix warning printf type specifier
solverd: fix format specifier warning
kcrumb-replay/ktest-bc: move unused variable aroudn to fix warning
expr/exprbuilder: ifdef out default case to fix warning
expr/exprvisitor: get rid of default case for warning, replace with assert
expr/lexer: remove default switch case to fix warning
kleaver: remove default switch case to fix warning
Expr/smtlib: fix unused yyinput warning

+++ tools/kmc-replay/SyscallsKTest.cpp
@@ -487,9 +487,11 @@ void SyscallsKTest::doLinuxThunks(SyscallParams& sp, int xlate_sysnr)
 			KREPLAY_SC "WRITING %d bytes to fd=%d :\"\n",
 			(int)sp.getArg(2), (int)sp.getArg(0));
 		bw = write(STDERR_FILENO, sp.getArgPtr(1), sp.getArg(2));
-		if (bw != sp.getArg(2)) {
-			fprintf(stderr, "Couldn't write %u bytes. Got %ld\n",
-				sp.getArg(2), bw);
+		if (bw != (ssize_t)sp.getArg(2)) {
+			fprintf(stderr,
+				"Couldn't write %lu bytes. Got %ld\n",
+				sp.getArg(2),
+				bw);
 		}
 		fprintf(stderr, "\n" KREPLAY_SC "\".\n");
